Module Extension.Syntax

A lightweight syntax for accessing configuration parameters.

Once this module is opened it is possible to access the parameter value using the infix notation, e.g., ctxt-->arch.

val (-->) : ctxt -> 'a Configuration.param -> 'a