Module Bap_main_event

type t = ..

type event = t = ..

val stream : t Bap_future.Std.stream

val send : t -> unit

val register_printer : (t -> string option) -> unit

register_printer f when event e is printed, f e must be None if f is not a subset of events, that is intended to be printed by an f. If it is Some str, then str is printed out.

If more than one printer returns Some thing for the same event, then the last registered has the precedence.

module Log : sig ... end

val pp : Stdlib.Format.formatter -> t -> unit