Module Extension.Error

An extensible set of possible errors

type t = error = ..

type t += Configuration

type t += Invalid of string

type t += Exit_requested of int

type t += Unknown_plugin of string

type t += Bug of exn * string

val pp : Stdlib.Format.formatter -> t -> unit

pp ppf err outputs a human readable description of err

val register_printer : (t -> string option) -> unit

register_printer to_string registers a printer for a subset of the errors.