Module Bap_main_event

Event subsystem.

The event subsystem is a way of communicating between different components of BAP.

type t = ..

type event = t = ..

val stream : t Bap_future.Std.stream

the global stream of events

val send : t -> unit

send event sends the event to the global stream of events stream.

val register_printer : (t -> string option) -> unit

register_printer f when the event e is printed, f e must be None if f is not a subset of events, that is intended to be printed by f. If it is Some str, then str is printed out.

If more than one printer returns Some thing for the same event, then the last registered has the precedence.

module Log : sig ... end

Logging Events

val pp : Stdlib.Format.formatter -> t -> unit

pp ppf event outputs event to the formatter ppf.